@@ -1,12 +1,20 @@
+buildscript {
+	ext {
+		queryDslVersion = "5.0.0"
+	}
+}
+
 plugins {
 	id 'org.springframework.boot' version '2.7.3'
 	id 'io.spring.dependency-management' version '1.0.13.RELEASE'
+	//querydsl 추가
+	id "com.ewerk.gradle.plugins.querydsl" version "1.0.10"
 	id 'java'
 }
 
 group = 'com.example'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '1.8'
+sourceCompatibility = '11'
 
 configurations {
 	compileOnly {
@@ -28,7 +36,8 @@ dependencies {
 	annotationProcessor 'org.projectlombok:lombok'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	implementation 'org.springframework.boot:spring-boot-starter-mustache'
-
+	// mysql
+	runtimeOnly 'mysql:mysql-connector-java'
 
 	//	소셜로그인
 	implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
@@ -52,15 +61,21 @@ dependencies {
 	//	JWT
 	compileOnly 'com.auth0:java-jwt:3.13.0'
 	implementation 'io.jsonwebtoken:jjwt:0.9.1'
+	implementation 'org.json:json:20200518'
 	implementation group: 'com.auth0', name: 'java-jwt', version: '3.10.3'
 	implementation group: 'org.javassist', name: 'javassist', version: '3.15.0-GA'
 	annotationProcessor "org.springframework.boot:spring-boot-configuration-processor"
 	implementation 'commons-codec:commons-codec:1.13'
 
-	//websocket
-	implementation 'org.springframework.boot:spring-boot-starter-websocket'
-	implementation 'org.webjars:sockjs-client:1.1.2'
-	implementation 'org.webjars:stomp-websocket:2.3.3-1'
+	//메일 발송
+	implementation 'org.springframework.boot:spring-boot-starter-mail'
+
+	//Redis TODO: redis(health check) warn나면 gradle 주석처리해주세요
+//	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+
+	//querydsl 추가
+	implementation "com.querydsl:querydsl-jpa:${queryDslVersion}"
+	annotationProcessor "com.querydsl:querydsl-apt:${queryDslVersion}"
 
 	//view
 
@@ -74,12 +89,46 @@ dependencies {
 	//s3
 	implementation 'org.springframework.cloud:spring-cloud-starter-aws:2.0.1.RELEASE'
 	implementation 'io.jsonwebtoken:jjwt-api:0.11.2'
-
+	implementation group: 'org.springframework.cloud', name: 'spring-cloud-aws', version: '2.2.6.RELEASE', ext: 'pom'
 	//jasypt
-	implementation 'com.github.ulisesbocchio:jasypt-spring-boot-starter:3.0.4'
+//	implementation 'com.github.ulisesbocchio:jasypt-spring-boot-starter:3.0.4'
+	implementation group: 'com.github.ulisesbocchio', name: 'jasypt-spring-boot-starter', version: '3.0.4'
+
+	// validation
+	implementation 'org.springframework.boot:spring-boot-starter-validation'
+
+	//prometheus
+	implementation 'io.micrometer:micrometer-registry-prometheus'
+	implementation 'org.springframework.boot:spring-boot-starter-actuator:2.7.2'
+	implementation 'io.micrometer:micrometer-core'
+
+	// Marvin
+	implementation 'com.github.downgoon:marvin:1.5.5'
+	implementation 'com.github.downgoon:MarvinPlugins:1.5.5'
+	compileOnly(group: 'org.springframework', name: 'spring-test')
 
 }
 
 tasks.named('test') {
 	useJUnitPlatform()
+}
+
+//querydsl 추가 시작
+def querydslDir = "$buildDir/generated/querydsl"
+
+querydsl {
+	jpa = true
+	querydslSourcesDir = querydslDir
+}
+sourceSets {
+	main.java.srcDir querydslDir
+}
+compileQuerydsl{
+	options.annotationProcessorPath = configurations.querydsl
+}
+configurations {
+	compileOnly {
+		extendsFrom annotationProcessor
+	}
+	querydsl.extendsFrom compileClasspath
 }
