@@ -1,15 +1,45 @@
 package com.example.checkcheck;
 
+import com.querydsl.jpa.impl.JPAQueryFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.context.annotation.Bean;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.scheduling.annotation.EnableAsync;
+
+import javax.annotation.PostConstruct;
+import javax.persistence.EntityManager;
+import java.util.TimeZone;
 
 @EnableJpaAuditing
 @SpringBootApplication
 public class StompApplication {
 
+	public static final String APPLICATION_LOCATIONS = "spring.config.location="
+			+ "classpath:application.yml,"
+			+ "/app/config/springboot-webservice/real-application.yml";
+
 	public static void main(String[] args) {
-		SpringApplication.run(StompApplication.class, args);
+		new SpringApplicationBuilder(StompApplication.class)
+				.properties(APPLICATION_LOCATIONS)
+				.run(args);
+
+	}
+
+//	public static void main(String[] args) {
+//		SpringApplication.run(StompApplication.class, args);
+//
+//	}
+
+	@PostConstruct
+	public void started(){
+		TimeZone.setDefault(TimeZone.getTimeZone("Asia/Seoul"));
+
+	}
+	@Bean
+	JPAQueryFactory jpaQueryFactory(EntityManager em) {
+		return new JPAQueryFactory(em);
 	}
 
 }
