@@ -1,13 +1,22 @@
 package com.example.checkcheck.controller;
 
 import com.example.checkcheck.dto.requestDto.ArticleRequestDto;
+import com.example.checkcheck.dto.responseDto.ArticleDetailResponseDto;
 import com.example.checkcheck.dto.responseDto.ArticleResponseDto;
+import com.example.checkcheck.dto.responseDto.ResponseDto;
+import com.example.checkcheck.dto.responseDto.StatusResponseDto;
+import com.example.checkcheck.model.articleModel.Category;
+import com.example.checkcheck.model.articleModel.Process;
+import com.example.checkcheck.repository.ArticleRepository;
+import com.example.checkcheck.security.UserDetailsImpl;
 import com.example.checkcheck.service.ArticleService;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestPart;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -18,14 +27,51 @@ import java.util.List;
 public class ArticleController {
 
     private ArticleService articleService;
+    private ArticleRepository articleRepository;
 
-    public ArticleController(ArticleService articleService) {
+    public ArticleController(ArticleService articleService, ArticleRepository articleRepository) {
         this.articleService = articleService;
+        this.articleRepository = articleRepository;
     }
 
-    @PostMapping(value = "/form", consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE})
-    public ArticleResponseDto upload(@RequestPart(required = false) ArticleRequestDto articleRequestDto,
-                                     @RequestPart (required = false) List<MultipartFile> multipartFile) throws IOException {
-        return articleService.postArticles(multipartFile, articleRequestDto);
+    @PostMapping(value = "/auth/form", consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE})
+    public ResponseDto<?> upload(@RequestPart(required = false) ArticleRequestDto articlesDto,
+                                 @RequestPart(required = false) List<MultipartFile> multipartFile,
+                                 @AuthenticationPrincipal UserDetailsImpl userDetails) throws IOException {
+        return articleService.postArticles(multipartFile, articlesDto, userDetails);
     }
+
+    @GetMapping("/main/list")
+    public Slice<ArticleResponseDto> getAllArticle(@RequestParam(value = "category", required = false) Category category,
+                                                   @RequestParam(value = "process", required = false) Process process,
+                                                   Pageable pageable) {
+
+        return articleService.getAllArticles(pageable, category, process);
+    }
+
+    @GetMapping("/main/randomcards")
+    public ResponseEntity<Object> getArticleCarousel() {
+        List<ArticleResponseDto> articleResponseDtos = articleService.getArticleCarousel();
+        return new ResponseEntity<>(new StatusResponseDto("Success", articleResponseDtos), HttpStatus.OK);
+    }
+
+    @GetMapping("/auth/detail/{articlesId}")
+    public ArticleDetailResponseDto getArticleDetail(@PathVariable Long articlesId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        return articleService.getArticleDetail(articlesId, userDetails);
+    }
+
+    @PatchMapping("/auth/detail/{articlesId}")
+    public ResponseDto<?> patchArticle(@RequestPart(required = false) ArticleRequestDto articlesDto,
+                                     @RequestPart(required = false) List<MultipartFile> multipartFile,
+                                     @PathVariable Long articlesId,
+                                     @AuthenticationPrincipal UserDetailsImpl userDetails) throws IOException {
+        return articleService.patchArticle(multipartFile, articlesDto, articlesId, userDetails);
+    }
+
+
+    @DeleteMapping("/auth/detail/{articlesId}")
+    public ResponseDto<?> deleteArticle(@PathVariable Long articlesId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+       return  articleService.deleteArticle(articlesId, userDetails);
+    }
+
 }
