@@ -9,13 +9,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class SimpleController {
 
-    @GetMapping("/lalalalala")
-    public String simpleCon(@AuthenticationPrincipal UserDetailsImpl userDetails) {
+    @GetMapping("/")
+    public String simpleCon() {
 
-        String name = SecurityContextHolder.getContext().getAuthentication().getName();
-        System.out.println("name = " + name);
-        System.out.println("userDetails username= " + userDetails.getUsername());
-        return userDetails.getUsername();
+        return "수정444";
 
     }
+    @GetMapping("/health")
+    public String checkHealth() {
+        return "healthy";
+    }
+
 }
