@@ -1,8 +1,10 @@
 package com.example.checkcheck.dto.requestDto;
 
-import com.example.checkcheck.model.Category;
+import com.example.checkcheck.model.articleModel.Category;
 import lombok.Getter;
 import lombok.Setter;
+
+import java.util.List;
 
 @Getter
 @Setter
@@ -12,4 +14,6 @@ public class ArticleRequestDto {
     private String content;
     private int price;
     private Category category;
+    private List<String> imageList;
+
 }
