@@ -4,17 +4,23 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Entity;
-import javax.persistence.Id;
+import javax.persistence.*;
 
 @Entity
 @Data
 @NoArgsConstructor
+@Table(name = "RefreshToken")
 public class RefreshToken {
 
     @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long tokenId;
+
+
+    @Column
     private String tokenKey;
 
+    @Column
     private String tokenValue;
 
     @Builder
