@@ -2,14 +2,16 @@ package com.example.checkcheck.repository;
 
 import com.example.checkcheck.model.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, Long>, CrudRepository<Member, Long> {
 
     Optional<Member> findByUserRealEmail(String email);
     Optional<Member> findByUserEmail(String email);
 
     Optional<Member> findByNickName(String email);
 
+    Optional<Member> findByUserName(String nickName);
 }
