@@ -1,7 +1,9 @@
 package com.example.checkcheck.security;
 
+import com.example.checkcheck.exception.CustomException;
+import com.example.checkcheck.exception.ErrorCode;
 import com.example.checkcheck.model.Member;
-import com.example.checkcheck.repository.UserRepository;
+import com.example.checkcheck.repository.MemberRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -12,14 +14,14 @@ import org.springframework.stereotype.Service;
 public class UserDetailsServiceImpl implements UserDetailsService {
 
     @Autowired
-    public UserDetailsServiceImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
+    public UserDetailsServiceImpl(MemberRepository memberRepository) {
+        this.memberRepository = memberRepository;
     }
 
-    private final UserRepository userRepository;
+    private final MemberRepository memberRepository;
 
     public UserDetails loadUserByUsername(String userEmail) throws UsernameNotFoundException {
-        Member member = userRepository.findByUserEmail(userEmail)
+        Member member = memberRepository.findByUserEmail(userEmail)
                 .orElseThrow(() -> new UsernameNotFoundException("Can't find " + userEmail));
 
         return new UserDetailsImpl(member);
