@@ -1,17 +1,21 @@
 package com.example.checkcheck.service.social;
 
 import com.example.checkcheck.dto.responseDto.SocialResponseDto;
+import com.example.checkcheck.dto.responseDto.TokenFactory;
 import com.example.checkcheck.dto.userinfo.NaverUserInfoDto;
 import com.example.checkcheck.model.Member;
 import com.example.checkcheck.model.RefreshToken;
 import com.example.checkcheck.repository.RefreshTokenRepository;
-import com.example.checkcheck.repository.UserRepository;
+import com.example.checkcheck.repository.MemberRepository;
 import com.example.checkcheck.security.UserDetailsImpl;
-import com.example.checkcheck.service.UserService;
+import com.example.checkcheck.service.MemberService;
+//import com.example.checkcheck.service.RedisService;
+import com.example.checkcheck.util.ComfortUtils;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
+//import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -24,6 +28,7 @@ import java.io.*;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.time.LocalDateTime;
+import java.util.Optional;
 import java.util.UUID;
 @RequiredArgsConstructor
 @Service
@@ -34,10 +39,13 @@ public class SocialNaverSerivce {
     private String client_id;
     @Value("${cloud.security.oauth2.client.registration.naver.client-secret}")
     private String clientSecret;
-    private final UserRepository userRepository;
+    private final MemberRepository memberRepository;
     private final PasswordEncoder passwordEncoder;
-    private final UserService userService;
+    private final MemberService memberService;
     private final RefreshTokenRepository refreshTokenRepository;
+    private final ComfortUtils comfortUtils;
+
+//    private final RedisService redisService;
 
     @Transactional
     public SocialResponseDto naverLogin(String code, String state, HttpServletResponse response) {
@@ -52,24 +60,25 @@ public class SocialNaverSerivce {
             String realEmail = "n_" + naverUserEmail;
 
 
+
             // 재가입 방지
             // 네이버 ID로 유저 정보 DB 에서 조회
-            Member member = userRepository.findByUserEmail(realEmail).orElse(null);
+            Member member = memberRepository.findByUserEmail(realEmail).orElse(null);
 
 
             // 없으면 회원가입
             if (member == null) {
                 member = Member.builder()
                         .nickName(naverUser.getNickname().substring(1, naverUser.getNickname().length() - 1))
+                        .userName(naverUser.getNickname().substring(1, naverUser.getNickname().length() - 1))
                         .password(encodedPassword)
                         .userEmail("n_"+naverUserEmail)
                         .userRealEmail(naverUser.getUserEmail().substring(1, naverUser.getUserEmail().length() - 1))
-                        .userRank("Bronze")
 //                        .socialId(Double.valueOf(naverUser.getNaverId().substring(1, naverUser.getNaverId().length() - 1)))
                         .createdAt(LocalDateTime.now())
                         .provider(provider)
                         .build();
-                userRepository.save(member);
+                memberRepository.save(member);
 
             } else {
                 // 강제 로그인
@@ -79,29 +88,36 @@ public class SocialNaverSerivce {
             }
 
 
-            System.out.println("결과는"+ member.getUserEmail());
             // User 권한 확인
 
 
 //            토큰 관리
-            String jwtToken = userService.accessAndRefreshTokenProcess(member.getUserEmail(), response);
+            TokenFactory tokenFactory = memberService.accessAndRefreshTokenProcess(member.getUserEmail(), response);
+
+            String refreshToken =  tokenFactory.getRefreshToken();
+
+//            redisService.setValues(member.getUserEmail(), refreshToken);
 
             String accessToken = naverUser.getAccessToken();
-            String refreshToken = naverUser.getRefreshToken();
 
 
-//            리프레시 토큰 저장
+//        리프레시토큰저장 & 있을경우 셋토큰
+//            TODO: Redis 실습
+            Optional<RefreshToken> existToken = refreshTokenRepository.findByTokenKey(member.getUserEmail());
             RefreshToken token = new RefreshToken(member.getUserEmail(), refreshToken);
-            refreshTokenRepository.save(token);
+            if (existToken.isEmpty()) {
+                refreshTokenRepository.save(token);
+            } else {
+                existToken.get().setTokenKey(token.getTokenKey());
+                existToken.get().setTokenValue(token.getTokenValue());
+            }
 
             SocialResponseDto socialResponseDto = SocialResponseDto.builder()
                     .nickName(member.getNickName()) // 1
                     .userEmail(member.getUserEmail())
-                    .accessToken(accessToken)
-                    .refreshToken(refreshToken)
-                    .userRank(member.getUserRank())
-                    .refreshToken(refreshToken)
-                    .jwtToken("Bearer "+jwtToken)
+                    .accessToken(tokenFactory.getAccessToken())
+                    .refreshToken(tokenFactory.getRefreshToken())
+                    .userRank(comfortUtils.getUserRank(member.getPoint()))
                     .build();
             return socialResponseDto;
 
@@ -128,7 +144,9 @@ public class SocialNaverSerivce {
             String sb = "grant_type=authorization_code" +
                     "&client_id="+client_id +
                     "&client_secret="+clientSecret +
-                    "&redirect_uri=http://localhost:8080/user/signin/naver" +
+//                    "&redirect_uri=http://localhost:8080/user/signin/naver" +
+//                    "&redirect_uri=http://localhost:3000/user/signin/naver" +
+                    "&redirect_uri=https://www.chackcheck99.com/signin/naver" +
                     "&code=" + code +
                     "&state=" + state;
             bw.write(sb);
